fix comment review_id
